update logic delete post

+++ src/hooks/useDeletePost.ts
@@ -5,7 +5,7 @@ import { useRevalidatePostList } from './useRevalidatePostList';
 import { GetPostsResponse } from './useGetPosts';
 
 interface Props {
-    mutate: KeyedMutator<GetPostsResponse | undefined>;
+    mutate?: KeyedMutator<GetPostsResponse | undefined>;
 }
 
 const useDeletePost = ({ mutate }: Props) => {
@@ -13,47 +13,41 @@ const useDeletePost = ({ mutate }: Props) => {
     const { enqueueSnackbar } = useSnackbar();
 
     const handleDeletePost = async (id: string) => {
-        const optimisticData = (currentData: GetPostsResponse | undefined) => {
-            if (!currentData) return undefined;
-            return {
-                ...currentData,
-                data: currentData.data.filter((it) => it.id !== id),
-            };
-        };
         try {
-            await mutate(
-                new Promise((resolve) => {
-                    setTimeout(() => {
-                        // const randomErr = Math.floor(Math.random() * 2);
-                        // if (randomErr === 1) {
-                        //     return reject(new Error('Something wrong'));
-                        // }
-                        instance.delete('/posts/' + id).then((data) => {
-                            enqueueSnackbar({
-                                variant: 'success',
-                                message: 'Delete ' + data.data.title + ' successfully',
-                            });
-                            resolve(data as any);
-                        });
-                    }, 500);
-                }),
-                {
-                    // During calling API
-                    optimisticData(_currentData, displayData) {
-                        return optimisticData(displayData);
-                    },
-                    populateCache: false,
-                    throwOnError: true,
-                    rollbackOnError: true,
-                    revalidate: false,
+            mutate?.(undefined, {
+                optimisticData(_currentData, displayedData) {
+                    if (!displayedData) return undefined;
+                    return {
+                        ...displayedData,
+                        data: displayedData.data.filter((it) => it.id !== id),
+                    };
                 },
-            );
-            revalidateAll();
+                populateCache: false,
+                revalidate: false,
+            });
+
+            await new Promise((resolve, _reject) => {
+                setTimeout(() => {
+                    const randomErr = Math.floor(Math.random() * 2);
+                    if (randomErr === 1 && !import.meta.env.TEST) {
+                        return _reject(new Error('Something wrong'));
+                    }
+                    instance.delete('/posts/' + id).then((data) => {
+                        enqueueSnackbar({
+                            variant: 'success',
+                            message: 'Delete ' + data.data.title + ' successfully',
+                        });
+                        resolve(data as any);
+                    });
+                }, 500);
+            });
         } catch (error) {
             enqueueSnackbar({
                 variant: 'error',
                 message: 'Fail to delete item',
             });
+        } finally {
+            revalidateAll();
         }
     };
